+++ pykafka/__init__.py
@@ -7,9 +7,11 @@ from .topic import Topic
 from .connection import SslConfig
 from .client import KafkaClient
 from .balancedconsumer import BalancedConsumer
+from .managedbalancedconsumer import ManagedBalancedConsumer
 
-__version__ = '2.2.2-dev'
+__version__ = '2.3.2-dev'
 
 
 __all__ = ["Broker", "SimpleConsumer", "Cluster", "Partition", "Producer",
-           "Topic", "SslConfig", "KafkaClient", "BalancedConsumer"]
+           "Topic", "SslConfig", "KafkaClient", "BalancedConsumer",
+           "ManagedBalancedConsumer"]