+++ pykafka/exceptions.py
@@ -159,18 +159,60 @@ class GroupLoadInProgress(ProtocolClientError):
     ERROR_CODE = 14
 
 
-class ConsumerCoordinatorNotAvailable(ProtocolClientError):
+class GroupCoordinatorNotAvailable(ProtocolClientError):
     """The broker returns this error code for consumer metadata requests or
         offset commit requests if the offsets topic has not yet been created.
     """
     ERROR_CODE = 15
 
 
-class NotCoordinatorForConsumer(ProtocolClientError):
+class NotCoordinatorForGroup(ProtocolClientError):
     """The broker returns this error code if it receives an offset fetch or
         commit request for a consumer group that it is not a coordinator for.
     """
     ERROR_CODE = 16
+
+
+class IllegalGeneration(ProtocolClientError):
+    """Returned from group membership requests (such as heartbeats) when the generation
+        id provided in the request is not the current generation
+    """
+    ERROR_CODE = 22
+
+
+class InconsistentGroupProtocol(ProtocolClientError):
+    """Returned in join group when the member provides a protocol type or set of protocols
+        which is not compatible with the current group.
+    """
+    ERROR_CODE = 23
+
+
+class UnknownMemberId(ProtocolClientError):
+    """Returned from group requests (offset commits/fetches, heartbeats, etc) when the
+        memberId is not in the current generation.
+    """
+    ERROR_CODE = 25
+
+
+class InvalidSessionTimeout(ProtocolClientError):
+    """Returned in join group when the requested session timeout is outside of the allowed
+        range on the broker
+    """
+    ERROR_CODE = 26
+
+
+class RebalanceInProgress(ProtocolClientError):
+    """Returned in heartbeat requests when the coordinator has begun rebalancing the
+        group. This indicates to the client that it should rejoin the group.
+    """
+    ERROR_CODE = 27
+
+
+class GroupAuthorizationFailed(ProtocolClientError):
+    """Returned by the broker when the client is not authorized to access a particular
+    groupId.
+    """
+    ERROR_CODE = 30
 
 
 ERROR_CODES = dict(
@@ -186,8 +228,14 @@ ERROR_CODES = dict(
                 MessageSizeTooLarge,
                 OffsetMetadataTooLarge,
                 GroupLoadInProgress,
-                ConsumerCoordinatorNotAvailable,
-                NotCoordinatorForConsumer)
+                GroupCoordinatorNotAvailable,
+                NotCoordinatorForGroup,
+                IllegalGeneration,
+                InconsistentGroupProtocol,
+                UnknownMemberId,
+                InvalidSessionTimeout,
+                RebalanceInProgress,
+                GroupAuthorizationFailed)
 )