+++ pykafka/rdkafka/producer.py
@@ -32,7 +32,8 @@ class RdKafkaProducer(Producer):
                  linger_ms=5 * 1000,
                  block_on_queue_full=True,
                  sync=False,
-                 delivery_reports=False):
+                 delivery_reports=False,
+                 auto_start=True):
         callargs = {k: v for k, v in vars().items()
                          if k not in ("self", "__class__")}
         self._rdk_producer = None
@@ -99,7 +100,6 @@ class RdKafkaProducer(Producer):
 
             # NB these refer not to payloads, but to wire messages
             # We've no real equivalents for these, but defaults should be fine:
-            ##"message.max.bytes"
             ##"receive.message.max.bytes"
 
             # No direct equivalents:
@@ -138,6 +138,7 @@ class RdKafkaProducer(Producer):
             "retry.backoff.ms": self._retry_backoff_ms,
             "compression.codec": map_compression_types[self._compression],
             "batch.num.messages": self._min_queued_messages,
+            "message.max.bytes": self._max_request_size,
 
             # Report successful and failed messages so we know to dealloc them
             "delivery.report.only.error": "false",